core-8737 Refactor PermissionValue. Address comments in PR#133
