+++ de-lib/src/main/java/org/iplantc/de/apps/client/events/tools/DeleteToolSelected.java
@@ -10,9 +10,9 @@ import com.google.gwt.event.shared.HandlerRegistration;
 import java.util.List;
 
 /**
- * Created by jstroot on 3/5/15.
  *
- * @author jstroot
+ *
+ * @author sriram
  */
 public class DeleteToolSelected extends GwtEvent<DeleteToolSelected.DeleteToolsSelectedHandler> {
     public static interface DeleteToolsSelectedHandler extends EventHandler {
