+++ de-lib/src/main/java/org/iplantc/de/diskResource/client/model/DataSharingProperties.java
@@ -3,7 +3,7 @@
  */
 package org.iplantc.de.diskResource.client.model;
 
-import org.iplantc.de.client.models.diskResources.PermissionValue;
+import org.iplantc.de.client.models.sharing.PermissionValue;
 import org.iplantc.de.client.models.sharing.Sharing;
 
 import com.sencha.gxt.core.client.ValueProvider;
