+++ de-lib/src/main/java/org/iplantc/de/diskResource/client/views/details/DetailsViewImpl.java
@@ -3,7 +3,7 @@ package org.iplantc.de.diskResource.client.views.details;
 import org.iplantc.de.client.models.diskResources.DiskResource;
 import org.iplantc.de.client.models.diskResources.File;
 import org.iplantc.de.client.models.diskResources.Folder;
-import org.iplantc.de.client.models.diskResources.PermissionValue;
+import org.iplantc.de.client.models.sharing.PermissionValue;
 import org.iplantc.de.client.models.search.DiskResourceQueryTemplate;
 import org.iplantc.de.client.models.search.SearchAutoBeanFactory;
 import org.iplantc.de.client.models.viewer.InfoType;
