+++ de-lib/src/test/java/org/iplantc/de/apps/client/views/toolBar/AppsViewToolbarImplTest.java
@@ -18,7 +18,7 @@ import org.iplantc.de.client.models.apps.AppAutoBeanFactory;
 import org.iplantc.de.client.models.apps.AppCategory;
 import org.iplantc.de.client.models.apps.proxy.AppLoadConfig;
 import org.iplantc.de.client.models.apps.proxy.AppSearchAutoBeanFactory;
-import org.iplantc.de.client.models.diskResources.PermissionValue;
+import org.iplantc.de.client.models.sharing.PermissionValue;
 import org.iplantc.de.client.services.AppServiceFacade;
 
 import com.google.common.collect.Lists;
